requirements.txt
# Azure AI Foundry SDK
azure-ai-inference>=1.0.0b9
azure-identity>=1.15.0

# Azure OpenAI SDK
openai>=1.0.0

# Microsoft Agent Framework
agent-framework>=0.1.0

# MCP Protocol
mcp>=1.0.0

# Image Processing
pytesseract==0.3.10
Pillow>=10.2.0
phi-3-vision-mlx>=0.0.3rc1

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.7.2
PyYAML>=6.0
requests>=2.31.0

# CLI & Terminal UI
rich>=13.7.0
click>=8.1.7

# Testing
pytest>=7.4.0
